+++ HISTORY.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Fix bug with serializing dictionaries.
 * Fix error raised when serializing empty list.
+* Add ``only`` and ``exclude`` parameters to Serializer constructor.
 
 0.3.0 (2013-11-14)
 ++++++++++++++++++